app/requirements.txt
fastapi
uvicorn
pandas
openpyxl
playwright
pymongo
bcrypt
PyJWT
python-dotenv
certifi
python-multipart
